programmatically searcbar problem
how to add SearchController’s SearchBar to UIView programmatically
here is a genius way from  below tutorial , it use a UiView created in
storyboard to add the SearchBar as its subview
http://stackoverflow.com/questions/26639791/position-uisearchbar-program
atically-using-uisearchcontroller
Problem is that SearchBar looks exceed the margin of its super UiView
even through  its constraint and size has been set in storyboard when
app initialized.
And it’s so weird that search bar’s appearance will become fit in its
super uiview after it typed and resume.