Interaction between searchResultsController and Original UiView
Create AutoCompleteController class for searchResultsController of
SearchController
Sync place id and place between AutoCompleteController and main UiView.
Show text on Label when it finish
searchResultsController is a parameter in SearchController.
I had tried to build a storyboard tableview for searchResultsController
to use segue to go back to main UiView when search ends. (tap a table
view cell to select a autocomplete result) ,but prepareSegue function
was not available at this way.
So , Another solution is more trickier , use func
didSelectRowAtIndexPath to sync a selectedPath variable then most
important is , deactivate SearchController to trigger main UiView’s fun
searchBarTextDidEndEditing , and pass name and place id of selected
result.